requirement.txt
python==3.10.9
pytorch==2.0.0
torchvision==0.15.0
robustness==1.2.1
osqp==0.6.0
gurobi==10.0.1
pip install git+https://github.com/sun-umn/PyGRANSO.git
pip install git+https://github.com/fra31/auto-attack

Note:
the robustness package souce code need to be modified (they are out-dated).

1) .......\robustness\imagenet_models\alexnet.py  
        Line 2: from torchvision.models.utils import load_state_dict_from_url ==> from torch.hub import load_state_dict_from_url
2) .......\robustness\imagenet_models\squeezenet.py
        same as 1)
3) .......\robustness/imagenet_models/vgg.py
        same as 1)